requirements.txt
datasets==2.14.5
python-dotenv==1.0.0
pytorch_lightning==1.9.5
torch==2.1.0
transformers==4.34.0
wandb==0.16.0
scikit-learn==1.3.1
ray==2.7.1
